+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+* Fix - the /tests directory was badly configured in the gitattributes and not actually excluded from the package.
+
 * Add a ConsistentStringScrambler - Consistently "randomise" the words in a string to be the same for the same random seed value using a Seeded Fisher-Yates shuffle
 
 * Add classes to simplify the creation and parsing of XML sitemaps